Refactoring and adding a src/ folder
